pyproject.toml
[tool.poetry]
name = "real-time-chat"
version = "0.1.0"
description = ""
authors = ["Gaurav Gosain <itsgauravgosain@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
coqui-tts = "^0.24.1"
realtimestt = "^0.2.2"
websockets = "^12.0"
colorama = "^0.4.6"
keyboard = "^0.13.5"
torch = "2.3.1"
torchaudio = "2.3.1"
groq = "^0.9.0"
uuid = "^1.30"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
